fix: resolve all 274 mypy strict-mode errors and remove ignore_errors exclusions
Systematically fixed every mypy type error across the entire codebase
under strict type checking settings (disallow_untyped_defs,
disallow_any_generics, disallow_untyped_calls, warn_return_any, etc.),
eliminating the need for ignore_errors=true blanket suppressions.

Error categories resolved:
- no-any-return (102): Added bool(), float(), str(), int() casts for
  functions returning Any where specific types were declared
- union-attr (26): Added isinstance() guards and cast() for union type
  narrowing
- type-arg (26): Added type parameters for generic types (ndarray,
  Queue, Future, Task, etc.)
- arg-type (23): Fixed function argument type mismatches
- assignment (20): Fixed variable type annotations and expression types
- attr-defined (17): Fixed attribute access on incorrect types
- var-annotated (11): Added explicit type annotations
- index (10): Added None guards before indexing Optional values
- no-untyped-def (9): Added return type annotations
- operator (8): Fixed operand types for arithmetic/comparison operators
- unused-ignore (5): Removed stale type: ignore comments
- return-value (5): Fixed return type mismatches
- no-untyped-call (4): Added type annotations to called functions
- override (3): Fixed method signature overrides
- comparison-overlap (2): Fixed non-overlapping equality checks
- ignore-without-code (2): Removed invalid type: ignore comments
- has-type (1): Added type annotation to variable

Modules fixed (61 files across 25+ packages):
- cognitive: formal_verification, reflexion, predictive_coding,
  multi_agent_coordination, hierarchical_planning, chain_of_hindsight,
  chain_of_thought
- automl: optimizer, search_space
- detectors: uiux_anomaly, enhanced_statistical, advanced_physics,
  spectral_vibration, geological/disaster_detectors
- ml: meta_learning, training, few_shot_learning, cross_domain_transfer,
  cortical_network
- distributed/scaling: raft_consensus, cluster, distributed_processor
- federated: privacy, server, client, federated_robust
- biometric: iris, fingerprint, voice, __init__
- integrations: cross_platform_hub, weather, cache, financial
- infrastructure: observability, streaming
- quantum_computing: executor, circuits, detector
- And 20+ more across core, harmonics, explainability, emergent, api,
  gui, metrics, models, cli

pyproject.toml: Removed all three ignore_errors=true override blocks
that previously suppressed errors across ~90% of the codebase. All
modules now undergo full strict type checking.

Result: 0 mypy errors across 431 source files under strict settings.

+++ src/omni_mercury_engine/biometric/voice_recognition.py
@@ -335,7 +335,7 @@ class PitchExtractor:
             return 0.0
 
         pitch = self._sample_rate / peak_idx
-        return pitch
+        return float(pitch)
 
 
 class EnergyExtractor:
@@ -624,7 +624,7 @@ class VoiceLivenessDetector:
         if ratio < expected_ratio / 10:
             return 0.3
 
-        return min(1.0, 0.5 + pitch_peak / (avg_level * 4))
+        return float(min(1.0, 0.5 + pitch_peak / (avg_level * 4)))
 
     def _analyze_channel(self, audio: np.ndarray, sample_rate: int) -> float:
         """Analyze recording channel characteristics."""
